Move plugins to job.StateCtx usage instead of job.CancelCh/job.PauseCh
